Rename National Report sidebar
