+++ src/containers/sidebars/national-report-sidebar/challenges-sidebar/component.jsx
