+++ src/containers/sidebars/national-report-sidebar/challenges-sidebar/styles.module.scss
