+++ src/containers/sidebars/national-report-sidebar/overview-sidebar/local-species-card/local-species-card-styles.module.scss
